pyproject.toml
[tool.poetry]
name = "streamlit-google-auth-example"
version = "0.1.0"
description = ""
authors = ["hideto0710 <inamura@kageboushi.app>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
streamlit = "^0.84.1"
watchdog = "^2.1.3"
google-cloud-bigquery = "^2.21.0"
google-auth = "^1.33.0"
google-auth-oauthlib = "^0.4.4"
bokeh = "^2.3.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
